change int to size_t for matrix sizes

+++ tasks/luzan_e_matrix_rows_sum/common/include/common.hpp
@@ -7,9 +7,9 @@
 
 namespace luzan_e_matrix_rows_sum {
 
-using InType = std::tuple<std::vector<int>, int, int>;  // matrix, height, width
+using InType = std::tuple<std::vector<int>, size_t, size_t>;  // matrix, height, width
 using OutType = std::vector<int>;                       // vec of sums, size = height
-using TestType = std::tuple<int, int>;                  // height & width
+using TestType = std::tuple<size_t, size_t>;                  // height & width
 using BaseTask = ppc::task::Task<InType, OutType>;
 
 }  // namespace luzan_e_matrix_rows_sum
